refacter: move kosmo.ClientMock to kubeutil pkg
