+++ pkg/kubeutil/client_mock.go
@@ -1,4 +1,4 @@
-package kosmo
+package kubeutil
 
 import (
 	"context"
